Keep connection around
We had the awful idea to keep all Bluetooth connection in the MainPage.
That means it got "disconnected" when navigating away,
but at the same time it does remain functionally connected and thus
cannot be reconnected.

We now keep this in App, which seems to work well.

Ideally we'd move away from the procedural UI setting and more of a
declarative approach, but this works fine.

Fixes #1