Make the project build an exe
The default WinUI3 configuration builds an exe,
but it cannot be ran.

Instead the only real way to install the resulting package is using the
MSIX machinery.
Now, this is all lovely and even has great UI,
but it hard requires code signing and I do not have the money for it.

So, we build it in self-contained mode.